convert loop time to integer for metadata

+++ src/main/java/team1403/robot/Robot.java
@@ -41,7 +41,7 @@ public class Robot extends LoggedRobot {
     Logger.recordMetadata("2025 Robot", 
       "Debug Mode: " + Constants.DEBUG_MODE +
       "\nSysID Enabled: " + Constants.ENABLE_SYSID +
-      "\nLoop Time: " + Constants.kLoopTime*1000 + " ms" +
+      "\nLoop Time: " + (int)(Constants.kLoopTime*1000) + " ms" +
       "\nGit Commit: " + BuildConstants.GIT_SHA +
       "\nGit Branch: " + BuildConstants.GIT_BRANCH +
       "\nGit Dirty: " + BuildConstants.DIRTY +
